feat(project-settings): working toward Project team saving
